Ensure backward compatibility when using custom method
Some api client do not encode the ":" character despite RFC 3986.
So that when a query having ":" in the last path segment is issued,
it is interpreted as a separator for custom method, so that the call
failed.
As a fix, when the last path segment is parameterized, and the actual
path of an api contains ":" in it last segment, we consider it as
a whole parameter before issuing a 404.
For example : having `path/{id}` as API path, the following calls are valid :
* path/sampleId
* path/sampleId:someValue ('someValue' will be considered as part of the id)